Smoothed out passage way with gap points
