+++ parafoil/bspline.py
@@ -1,0 +1,15 @@
+from scipy.interpolate import BSpline
+import numpy as np
+import numpy.typing as npt
+
+def get_bspline(ctrl_pnts: npt.NDArray, degree: int):
+    "get a bspline with clamped knots"
+    num_ctrl_pnts = ctrl_pnts.shape[0]
+    knots = np.pad(
+        array=np.linspace(0, 1, (num_ctrl_pnts + 1) - degree),
+        pad_width=(degree, degree),
+        mode='constant',
+        constant_values=(0, 1)
+    )
+    return BSpline(knots, ctrl_pnts, degree, extrapolate=False)
+
